Moving away from LocalDateTime, in order to have a Generic overarching extension for all Java 8 DateTypes.
