--- jr-extension-datetime/src/moditect/module-info.java
@@ -1,6 +1,0 @@
-module com.fasterxml.jackson.jr.extension.datetime {
-    requires com.fasterxml.jackson.core;
-    requires com.fasterxml.jackson.jr.ob;
-
-    exports com.fasterxml.jackson.jr.extension.datetime;
-}
