+++ jr-extension-javatime/README.md
@@ -2,18 +2,24 @@
 
 This module extends the functionalities of jackson-jr adding support for a subset of Java 8 Date & Time objects such as `LocalDateTime` & `DateTime`.
 
-### Example
+### Status
+Added in Jackson 2.17.
+
+### Usage
+To be able to use supported annotations, you need to register extension like so:
+```java
+    private static final JSON JACKSON = JSON.builder()
+        .register(new JacksonLocalDateTimeExtension())
+        .build();
+```
+after which you can use normal read and write operations as usual:
 
 ```java
-import com.fasterxml.jackson.jr.extension.datetime.JacksonLocalDateTimeExtension;
-import com.fasterxml.jackson.jr.ob.JSON;
+
 
 import java.time.LocalDateTime;
 
 class Application {
-    private static final JSON JACKSON = JSON.builder()
-            .register(new JacksonLocalDateTimeExtension())
-            .build();
 
     public static void main(String[] args) {
         // Create the class that we need to
