+++ jr-extension-javatime/src/main/java/com/fasterxml/jackson/jr/extension/javatime/JacksonJavaTimeExtension.java
@@ -1,0 +1,11 @@
+package com.fasterxml.jackson.jr.extension.javatime;
+
+import com.fasterxml.jackson.jr.ob.JacksonJrExtension;
+import com.fasterxml.jackson.jr.ob.api.ExtensionContext;
+
+public class JacksonJavaTimeExtension extends JacksonJrExtension {
+    @Override
+    protected void register(ExtensionContext ctxt) {
+        ctxt.insertProvider(new JacksonJavaTimeReaderWriterProvider());
+    }
+}
