+++ jr-extension-javatime/src/main/java/com/fasterxml/jackson/jr/extension/javatime/JacksonJavaTimeValueReader.java
@@ -1,4 +1,4 @@
-package com.fasterxml.jackson.jr.extension.datetime;
+package com.fasterxml.jackson.jr.extension.javatime;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.jr.ob.api.ValueReader;
@@ -8,13 +8,16 @@ import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class LocalDateTimeValueReader extends ValueReader {
-    public LocalDateTimeValueReader() {
+public class JacksonJavaTimeValueReader extends ValueReader {
+    private final DateTimeFormatter formatter;
+
+    public JacksonJavaTimeValueReader(DateTimeFormatter formatter) {
         super(LocalDateTime.class);
+        this.formatter = formatter;
     }
 
     @Override
     public Object read(JSONReader reader, JsonParser p) throws IOException {
-        return LocalDateTime.parse(p.getText(), DateTimeFormatter.ISO_LOCAL_DATE_TIME);
+        return LocalDateTime.parse(p.getText(), formatter);
     }
 }
