+++ jr-extension-javatime/src/main/java/com/fasterxml/jackson/jr/extension/javatime/JacksonJavaTimeValueWriter.java
@@ -1,4 +1,4 @@
-package com.fasterxml.jackson.jr.extension.datetime;
+package com.fasterxml.jackson.jr.extension.javatime;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.jr.ob.api.ValueWriter;
@@ -8,10 +8,16 @@ import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class LocalDateTimeValueWriter implements ValueWriter {
+public class JacksonJavaTimeValueWriter implements ValueWriter {
+    private final DateTimeFormatter formatter;
+
+    public JacksonJavaTimeValueWriter(DateTimeFormatter formatter) {
+        this.formatter = formatter;
+    }
+
     @Override
     public void writeValue(JSONWriter context, JsonGenerator g, Object value) throws IOException {
-        String localDateTimeString = ((LocalDateTime) value).format(DateTimeFormatter.ISO_LOCAL_DATE_TIME);
+        String localDateTimeString = ((LocalDateTime) value).format(formatter);
         context.writeValue(localDateTimeString);
     }
