+++ jr-extension-javatime/src/main/resources/META-INF/LICENSE
