+++ jr-extension-javatime/src/moditect/module-info.java
@@ -1,0 +1,6 @@
+module com.fasterxml.jackson.jr.extension.javatime {
+    requires com.fasterxml.jackson.core;
+    requires com.fasterxml.jackson.jr.ob;
+
+    exports com.fasterxml.jackson.jr.extension.javatime;
+}
