+++ pom.xml
@@ -25,7 +25,7 @@
     <module>jr-retrofit2</module>
     <module>jr-stree</module>
     <module>jr-annotation-support</module>
-    <module>jr-extension-datetime</module>
+    <module>jr-extension-javatime</module>
     <module>jr-all</module>
   </modules>
